Enhance player insights components and introduce data context
- Refactored player insights components to utilize a new data context for improved data management and accessibility.
- Added new tabs for player insights, including games, overview, and people, enhancing the user experience with detailed statistics.
- Updated hooks and queries to align with the new data structure, ensuring consistency and maintainability.
- Improved error handling and logging for image cleanup processes in the EditPlayerDialog.
- Streamlined player dropdown mutation logic for better performance and clarity.

+++ packages/api/src/routers/match/sub-routers/update-match/update-match.input.ts
@@ -102,7 +102,7 @@ export const updateMatchPlacementsInput = z.object({
     .array(
       z.object({
         id: z.number(),
-        placement: z.number().min(1).nullable(),
+        placement: z.number().int().min(1).nullable(),
       }),
     )
     .refine((placements) => placements.length > 0),
